Augmented SoundCloud alert
In my tests, being on the song page in SoundCloud isn't always enough – it depends on the route by which I arrive on the song page and if I have played anything previously. So I've added a note that refreshing the song page may help, which it does in nearly every test I've run.